refactor(types): HomeTemplateProps を types/ へ集約しレイヤ逆流を解消する
useHomeScreen（hooks）が HomeTemplate（components）の型を import しており、
frontend.md「レイヤ依存の一方向ルール」に違反していた。

- HomeTemplateProps を types/home-template.ts へ移動し、hooks と components の
  双方がそこを参照する形にする（ルールの「逆流は共通化で解決する」表に従う）
- 36 メンバーすべてにコメントを付与。src/types/** は require-jsdoc が error
  （PR #150）のため必須。null や 0 件の意味・1 始まりの範囲・デバウンスの有無など、
  型シグネチャから読めない契約を記述した
- ValidationErrors も types/validation.ts へ昇格。4 ファイルから参照されており
  typescript.md「2 ファイル以上から参照される型は types/ へ集約」に該当する。
  加えて types/home-template.ts から lib/ を参照すると types/ が上位へ依存し、
  新たな逆流を作ってしまうため移動が必要だった
- 昇格に伴い元ファイルから型定義を削除（re-export も残さない）
- 型のみになった import を import type へ変更し、未使用 import を削除

Closes #139
